Partial fix for #15
For currently unknown reasons the quantity of some coins in the balance
command are 0 but the calculated value in btc is larger than 0. To get
those coins at least listed within the balance command the if condition
to check wheter a coin is included or not was adapted. Leaving back the
issue of the empty quantity.