Check Errors reported from Exchange and raise ApiError if there are any.
As reported in #33 the return value of a call of the Api is not checked. Now a
very simple _check_response method is implemented which looks for any error
messages in the response. If there is a error an ApiError is raised an the
excecution of the Program is aborted.